% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_config}
\alias{get_config}
\title{Query a configuration parameter}
\usage{
get_config(key)
}
\arguments{
\item{key}{The name of the parameter to query.}
}
\value{
The value of the parameter
}
\description{
This is the order in which the parameter is searched for: \enumerate{
  \item The configuration of the current session.
  \item The project configuration file in the current working
     directory.
  \item The user's configuration file.
  \item The site-wide configuration file.
}
}

