% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_config}
\alias{set_config}
\title{Set a configuration parameter}
\usage{
set_config(..., .where = c("user", "project", "session", "site"))
}
\arguments{
\item{...}{Parameters to set, they should be all named.}

\item{.where}{Where to set the parameters, \sQuote{user} is the
default and it means the user level configuration file.
\sQuote{project} means the project configuration file,
in the current directory. \sQuote{session} is not persistent,
only sets the parameter in a current R session. Finally, \sQuote{site}
means the site-level configuration file.}
}
\value{
Nothing.
}
\description{
Set a configuration parameter
}

