% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{set_config_in}
\alias{set_config_in}
\title{Set a configuration parameter for a package}
\usage{
set_config_in(..., .in = parent.frame())
}
\arguments{
\item{...}{Parameters to set, they should be all named.}

\item{.in}{An environment, typically belonging to a package.}
}
\value{
Nothing.
}
\description{
This is a more flexible variant of \code{\link{set_config}},
and it allows creating an custom API in the package that
uses \code{pkgconfig} for its configuration.
}
\details{
This function is identical to \code{\link{set_config}}, but it allows
supplying the environment that is used as the package the configuration
is set for. This makes it possible to create an API for setting
(and getting) configuration parameters.
}
\seealso{
\code{\link{set_config}}
}
