% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getset.R
\name{get_config}
\alias{get_config}
\title{Query a configuration parameter key}
\usage{
get_config(key, fallback = NULL)
}
\arguments{
\item{key}{The name of the parameter to query.}

\item{fallback}{Fallback if the parameter id not found anywhere.}
}
\value{
The value of the parameter, or the fallback value if not found.
}
\description{
Query a configuration parameter key
}

