% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getset.R
\name{set_config}
\alias{set_config}
\title{Set a configuration parameter}
\usage{
set_config(...)
}
\arguments{
\item{...}{Parameters to set, they should be all named.}
}
\value{
Nothing.
}
\description{
Set a configuration parameter, for the package we are calling from.
If called from the R prompt and not from a package, then it sets
the parameter for global environment.
}
\seealso{
\code{\link{set_config_in}}
}
